requirements.txt
functions-framework==3.3.0
pandas>=2.1.2,<=2.2.3
scikit-learn>=1.3.1,<=1.6.1
numpy==1.26.4
google-cloud-storage==2.10.0
google-api-core==2.17.1
python-dateutil==2.8.2
requests==2.31.0
tqdm>=4.63.0,<=4.67.1
configspace==1.2.1
python-dotenv==1.0.1
flask==2.3.3
httpx>=0.25.0,<=0.27.2
omegaconf>=2.1.2,<=2.3.0
password-strength>=0.0.3.post2,<=0.0.3.post2
cityhash>=0.4.3,<=0.4.7
sseclient-py>=1.8.0,<=1.8.0
tabpfn-client==0.1.4 